+++ mullvad-cli/src/cmds/mod.rs
@@ -25,6 +25,9 @@ pub use self::disconnect::Disconnect;
 mod lan;
 pub use self::lan::Lan;
 
+mod custom_dns;
+pub use self::custom_dns::CustomDns;
+
 mod reconnect;
 pub use self::reconnect::Reconnect;
 
@@ -60,6 +63,8 @@ pub fn get_commands() -> HashMap<&'static str, Box<dyn Command>> {
         Box::new(Disconnect),
         Box::new(Reconnect),
         Box::new(Lan),
+        #[cfg(windows)]
+        Box::new(CustomDns),
         Box::new(Relay),
         Box::new(Reset),
         #[cfg(target_os = "linux")]