+++ mullvad-daemon/src/settings.rs
@@ -3,6 +3,8 @@ use mullvad_types::{
     relay_constraints::{BridgeSettings, BridgeState, RelaySettingsUpdate},
     settings::Settings,
 };
+#[cfg(windows)]
+use std::net::IpAddr;
 use std::{
     fs::{self, File},
     io,
@@ -206,6 +208,15 @@ impl SettingsPersister {
         let should_save = Self::update_field(
             &mut self.settings.tunnel_options.generic.enable_ipv6,
             enable_ipv6,
+        );
+        self.update(should_save)
+    }
+
+    #[cfg(windows)]
+    pub fn set_custom_dns(&mut self, servers: Option<Vec<IpAddr>>) -> Result<bool, Error> {
+        let should_save = Self::update_field(
+            &mut self.settings.tunnel_options.generic.custom_dns,
+            servers,
         );
         self.update(should_save)
     }