+++ mullvad-types/src/settings/mod.rs
@@ -177,6 +177,8 @@ impl Default for TunnelOptions {
             generic: GenericTunnelOptions {
                 // Enable IPv6 be default on Android
                 enable_ipv6: cfg!(target_os = "android"),
+                #[cfg(windows)]
+                custom_dns: None,
             },
         }
     }