+++ talpid-core/src/tunnel_state_machine/connecting_state.rs
@@ -227,6 +227,11 @@ impl ConnectingState {
                     }
                 }
             }
+            #[cfg(windows)]
+            Some(TunnelCommand::CustomDns(servers)) => {
+                shared_values.custom_dns = servers;
+                SameState(self.into())
+            }
             Some(TunnelCommand::BlockWhenDisconnected(block_when_disconnected)) => {
                 shared_values.block_when_disconnected = block_when_disconnected;
                 SameState(self.into())