+++ talpid-core/src/tunnel_state_machine/disconnected_state.rs
@@ -82,6 +82,11 @@ impl TunnelState for DisconnectedState {
                 }
                 SameState(self.into())
             }
+            #[cfg(windows)]
+            Some(TunnelCommand::CustomDns(servers)) => {
+                shared_values.custom_dns = servers;
+                SameState(self.into())
+            }
             Some(TunnelCommand::BlockWhenDisconnected(block_when_disconnected)) => {
                 if shared_values.block_when_disconnected != block_when_disconnected {
                     shared_values.block_when_disconnected = block_when_disconnected;